Trying to improve onvif config
