+++ src/ros2/src/bob_monitor/README.md
@@ -5,7 +5,12 @@
 ## Nodes:
 
 - Prometheus Node
+- ONVIF Service Node
 
 ### Prometheus Node
 
 Used to provide and endpoint for Prometheus to call. It is responsible for shaping the data ready for ingress into Prometheus.
+
+### ONVIF Service Node
+
+Used to get camera details using the ONVIF protocol.
